+++ tests/ColumnSchemaTest.php
@@ -4,16 +4,24 @@ declare(strict_types=1);
 
 namespace Yiisoft\Db\Mssql\Tests;
 
-use PHPUnit\Framework\TestCase;
+use PDO;
+use Yiisoft\Db\Command\Param;
+use Yiisoft\Db\Expression\Expression;
+use Yiisoft\Db\Mssql\Column\BinaryColumnSchema;
 use Yiisoft\Db\Mssql\Tests\Support\TestTrait;
 use Yiisoft\Db\Query\Query;
+use Yiisoft\Db\Schema\Column\BooleanColumnSchema;
+use Yiisoft\Db\Schema\Column\DoubleColumnSchema;
+use Yiisoft\Db\Schema\Column\IntegerColumnSchema;
+use Yiisoft\Db\Schema\Column\StringColumnSchema;
+use Yiisoft\Db\Tests\Common\CommonColumnSchemaTest;
 
 use function str_repeat;
 
 /**
  * @group mssql
  */
-final class ColumnSchemaTest extends TestCase
+final class ColumnSchemaTest extends CommonColumnSchemaTest
 {
     use TestTrait;
 
@@ -59,5 +67,45 @@ final class ColumnSchemaTest extends TestCase
         $this->assertFalse($boolColPhpType);
 
         $db->close();
+    }
+
+    public function testColumnSchemaInstance()
+    {
+        $db = $this->getConnection(true);
+        $schema = $db->getSchema();
+        $tableSchema = $schema->getTableSchema('type');
+
+        $this->assertInstanceOf(IntegerColumnSchema::class, $tableSchema->getColumn('int_col'));
+        $this->assertInstanceOf(StringColumnSchema::class, $tableSchema->getColumn('char_col'));
+        $this->assertInstanceOf(DoubleColumnSchema::class, $tableSchema->getColumn('float_col'));
+        $this->assertInstanceOf(BinaryColumnSchema::class, $tableSchema->getColumn('blob_col'));
+        $this->assertInstanceOf(BooleanColumnSchema::class, $tableSchema->getColumn('bool_col'));
+    }
+
+    /** @dataProvider \Yiisoft\Db\Mssql\Tests\Provider\ColumnSchemaProvider::predefinedTypes */
+    public function testPredefinedType(string $className, string $type, string $phpType)
+    {
+        parent::testPredefinedType($className, $type, $phpType);
+    }
+
+    /** @dataProvider \Yiisoft\Db\Mssql\Tests\Provider\ColumnSchemaProvider::dbTypecastColumns */
+    public function testDbTypecastColumns(string $className, array $values)
+    {
+        parent::testDbTypecastColumns($className, $values);
+    }
+
+    public function testBinaryColumnSchema()
+    {
+        $binaryCol = new BinaryColumnSchema();
+        $binaryCol->dbType('varbinary');
+
+        $this->assertEquals(
+            new Expression('CONVERT(VARBINARY(MAX), 0x101112)'),
+            $binaryCol->dbTypecast("\x10\x11\x12"),
+        );
+        $this->assertEquals(
+            new Expression('CONVERT(VARBINARY(MAX), 0x101112)'),
+            $binaryCol->dbTypecast(new Param("\x10\x11\x12", PDO::PARAM_LOB)),
+        );
     }
 }